+++ test/base/src/androidTest/kotlin/io/realm/kotlin/test/shared/NullabilityTests.kt
@@ -116,7 +116,7 @@ class NullabilityTests {
                 testProperty(Nullability::objectIdField, ObjectId.from("507f191e810c19729de860ea"))
                 testProperty(Nullability::uuidField, RealmUUID.random())
                 testProperty(Nullability::binaryField, byteArrayOf(42))
-                testProperty(Nullability::mutableRealmIntField, MutableRealmInt.of(42))
+                testProperty(Nullability::mutableRealmIntField, MutableRealmInt.create(42))
                 // Manually removing RealmObject as nullableFieldTypes is not referencing the
                 // explicit subtype (Nullability). Don't know how to make the linkage without
                 // so it also works on Native.