+++ test/base/src/androidTest/kotlin/io/realm/kotlin/test/shared/dynamic/DynamicRealmObjectTests.kt
@@ -517,7 +517,7 @@ class DynamicRealmObjectTests {
 
         realm.close()
 
-        assertFailsWithMessage<IllegalStateException>(RealmObjectReference.INVALID_OBJECT) {
+        assertFailsWithMessage<IllegalStateException>(RealmObjectReference.INVALID_OBJECT_MSG) {
             first.getValue<DynamicRealmObject>("stringField")
         }
     }